[MINOR][PYTHON][TESTS] Move test `test_named_arguments_negative` to `test_arrow_python_udf`
### What changes were proposed in this pull request?
Move test `test_named_arguments_negative` to `test_arrow_python_udf`

### Why are the changes needed?
it seems was added in a wrong place, it only runs in Spark Connect, not Spark Classic.
After this PR, it will also be run in Spark Classic

### Does this PR introduce _any_ user-facing change?
no, test only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #46544 from zhengruifeng/move_test_named_arguments_negative.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
